Refs #17979 Added path to blob in documents manage_main view
Note that only manage_main view of Document will show the path.
manage_document view will not.
User needs 'View management screens' permission in order to access this feature.
If the file is not found, 'file not found' will be shown istead

This will require an update script to be run on products using this.
(bdr, bdr-test, cdr, cdr-test)